Move data versioning code to its own file
