+++ .gitignore
@@ -3,7 +3,7 @@ results/*.csv
 results/*.json
 results/test/*.csv
 results/test/*.json
-results/test.*/county
+results/test.*
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
